app/requirements.txt
# coding=utf8
Flask
flask-cors
pyMySQL
python-gitlab
pyartifactory
python-jenkins
flask-ldap3-login
flask-sqlalchemy
gevent
tencentcloud-sdk-python
tencentcloud-sdk-python-sms
gunicorn==19.7.1
gevent==22.10.2